Move voucher details dialog paths to qs
