+++ saleor/static/dashboard-next/discounts/urls.ts
@@ -3,6 +3,7 @@ import * as urlJoin from "url-join";
 
 import { ActiveTab, Dialog, Pagination } from "../types";
 import { SaleDetailsPageTab } from "./components/SaleDetailsPage";
+import { VoucherDetailsPageTab } from "./components/VoucherDetailsPage";
 
 export const discountSection = "/discounts/";
 
@@ -27,6 +28,16 @@ export const voucherSection = urlJoin(discountSection, "vouchers");
 export const voucherListPath = voucherSection;
 export const voucherListUrl = voucherListPath;
 export const voucherPath = (id: string) => urlJoin(voucherSection, id);
-export const voucherUrl = (id: string) => voucherPath(encodeURIComponent(id));
+export type VoucherUrlDialog =
+  | "assign-category"
+  | "assign-collection"
+  | "assign-country"
+  | "assign-product"
+  | "remove";
+export type VoucherUrlQueryParams = Pagination &
+  Dialog<VoucherUrlDialog> &
+  ActiveTab<VoucherDetailsPageTab>;
+export const voucherUrl = (id: string, params?: VoucherUrlQueryParams) =>
+  voucherPath(encodeURIComponent(id)) + "?" + stringifyQs(params);
 export const voucherAddPath = urlJoin(voucherSection, "add");
 export const voucherAddUrl = voucherAddPath;
